ZMPO_5 modyfikacja

+++ semestr3/ZMPO/ZMPO_4/KnapsackProblem.cpp
@@ -3,7 +3,7 @@
 template <class T>
 KnapsackProblem<T>::~KnapsackProblem()
 {
-	for (int i = 0; i < items.size(); i++)
+	for (size_t i = 0; i < items.size(); i++)
 		delete items[i];
 }
